+++ .old-configs/debug-jest-config.cjs
@@ -1,5 +1,0 @@
-const { getJestConfig } = require('@storybook/test-runner');
-
-const config = getJestConfig();
-console.log('Jest Config from test-runner:');
-console.log(JSON.stringify(config, null, 2));